mainly eliminate secret stored data used at load
Before, this whenever quoted data or just data needs to be used, it was
the exact data from the reader, and it was just stored in a secret
table. This works, but the tradeoff is that I can't save the Lua code to
be compiled later or used at bytecode. The Lua code does not have the quoted data.

So to solve this I now always print out the data. At load time this gets
converted to the same data again. Metadata is lost in the process but it
shouldn't be that important.

Also lots of rework to:
* move stuff to Compiler namespace
* consolidate most things to "state" variable which will be helpful when
* I break up the XBOX huge if elseif statement next

Also try shortening the safe lua variable conversion
make (def) return nil like Clojure
more tests